FIX: Error from format chooser and encryption functions[t23960][t23961][t23991]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@15619 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ plugins/format_chooser/pages/convert.php
@@ -51,7 +51,9 @@ convertImage($resource, $page, $alternative, $target, $width, $height, $profile)
 daily_stat('Resource download', $ref);
 resource_log($ref, LOG_CODE_DOWNLOADED, 0,$lang['format_chooser'], '',  $size);
 
-sendFile($target);
-unlink($target);
-
+if(file_exists($target))
+    {
+    sendFile($target);
+    unlink($target);
+    }
 ?>
